requirements.txt
Django==2.2.6

django-model-utils==3.1.2
Pillow==5.1.0
django-extensions==2.0.7

psycopg2==2.7.5
django-taggit==1.1.0

django-ckeditor==5.7.1
django-widget-tweaks==1.4.5
django-redis-sessions==0.6.1
requests==2.22.0

djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
drf-yasg==1.17.1